v34
Aight, update for yah all.  Sekhmet now kicks you in the balls if you get too close :D
Grapple now works and fires out a rope before dragging you in, sadly it doesn't have the swinging that I originally wanted but it's kind of became a sacrifice to avoid rewriting the character controller.  I've mentioned some instructions in #code for setting up grappleable places.  The Grappling Hook script is attached to the Mummy and the max grapple distance and speed can be tweaked from there.